HUD System Rework WIP
This update saves progress on the HUD System rework. It is attempting to simplify and give more control to the HUD, so that it may decide itself on how to retrieve necessary variables for drawing rather than rely on gamemode-defined strict components.

A HUD file is now a Lua file that returns a table that may contain functions. The three main interactions come from the keyword name of the function, which will be implemented on the HUD if function name is prefixed with this keyword followed by an underscore. These are:

Paint_: Makes this function run in GM:HUDPaint()
Panel_: This function returns a panel which will be part of the HUD (added and removed as HUD is turned on/off). Panels can be retrived in runtime in the HUD.Panels table on the HUD object.
Hook_: This function will be hooked to the hook of the name that comes after the prefix. For example "HUD:Hook_Think()" will run on Think. These hooks are added and removed as the HUD is enabled/disabled.

This gives much more control to HUDs, but on the other hand also require them to handle their own data retrieval rather than getting pre-defined values from the gamemode based on keyname. The data which was previously accessible this way but is local to another file will be exposed in public getters in a future update as the system develops.