Target ID
Reworks the way Target ID is structured, with relevancy to the HUD and hooks. Instead of a predefined set of enums, type is done by string, which could be any. Whenever a string type is used, it will run the HUD function HUD:DrawTargetID[type]() where [type] is replaced with this string. For example HUD:DrawTargetIDBuy() for buy-types. HUD:DrawTargetID() is a generic one that will be run if the specific type isn't implemented, or no type is provided (just a full string of text to display)

+ Added hook nzu_GetTargetIDText
+ Added hook nzu_TranslateTargetIDText which is run when the HUD doesn't implement the type to translate the type into a full string that can be passed to the generic function instead
+ Added Target ID functions to the Unlimited HUD
* Updated all Target ID code to this system
* Moved hudmanagement.lua into gamemode folder